Better web albums; Cleanup

+++ models/album_asset.go
@@ -1,7 +1,7 @@
 package models
 
 type AlbumAsset struct {
-	CreatedAt int    `gorm:"index:album_order,priority:2"`
+	CreatedAt int64  `gorm:"index:album_order,priority:2"`
 	AlbumID   uint64 `gorm:"primaryKey;index:album_order,priority:1"`
 	Album     Album  `gorm:"constraint:OnUpdate:CASCADE,OnDelete:CASCADE;"`
 	AssetID   uint64 `gorm:"primaryKey"`
