+++ models/group.go
@@ -2,8 +2,8 @@ package models
 
 type Group struct {
 	ID          uint64 `gorm:"primaryKey"`
-	CreatedAt   int
-	UpdatedAt   int
+	CreatedAt   int64
+	UpdatedAt   int64
 	CreatedByID uint64
 	CreatedBy   User   `gorm:"constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
 	Name        string `gorm:"type:varchar(300)"`
